+++ src/options.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 /// Options for passing to `cargo expand`/`cargo check`.
 #[derive(Clone, Default, Debug)]
-pub struct Options {
+pub(crate) struct Options {
     // Additional arguments to pass to `cargo expand`/`cargo check`.
     pub args: Vec<String>,
     // Additional env variables to pass to `cargo expand`/`cargo check`.