+++ CHANGELOG.md
@@ -24,7 +24,11 @@ Please make sure to add your changes to the appropriate categories:
 
 ### Changed
 
-- n/a
+- Passing no file patterns is now considered a failure:
+  - Calling `expand()` with an empty list of file patterns will fail.
+  - Calling `expand_opts()` with an empty list of file patterns will fail.
+  - Calling `expand_fail()` with an empty list of file patterns will fail.
+  - Calling `expand_opts_fail()` with an empty list of file patterns will fail.
 
 ### Deprecated