+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Please make sure to add your changes to the appropriate categories:
 ### Fixed
 
 - External dependencies of the crate are now properly mirrored by the test projects.
+- Features of the crate are now properly mirrored by the test projects.
 
 ### Performance